archive-chat-agent-api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pydantic
pydantic-settings
python-multipart
azure-storage-blob
azure-ai-documentintelligence
openai
tiktoken
azure-search-documents
azure-cosmos
azure-identity